# Improve producer ID expiration performance