@jolshan 2024-05-28
Thanks Calvin -- just to be clear, we are also changing the benchmark to not include loading all the producer IDs as well?