@CalvinConfluent 2024-05-29
@jolshan Yes, the loading part is excluded. I guess there is not much performance concern for the loading.